- RaidBattle 消息推送优化

+++ game-common/src/main/java/com/nekonade/common/dto/RaidBattleDTO.java
@@ -65,7 +65,7 @@ public class RaidBattleDTO {
 
         private Integer level = 1;
 
-        private long contributePoint;
+        private int contributePoint;
 
         private int turn;
